refactor(server): :recycle: refactored to result pattern in domain, application and graphql projects

+++ libs/server/core/application/Features/Publishers/Commands/DeletePublisherCommandHandler.cs
@@ -1,15 +1,15 @@
-using Kathanika.Core.Domain.Exceptions;
-
 namespace Kathanika.Core.Application.Features.Publishers.Commands;
 
-internal sealed class DeletePublisherCommandHandler(IPublisherRepository publisherRepository) : IRequestHandler<DeletePublisherCommand>
+internal sealed class DeletePublisherCommandHandler(IPublisherRepository publisherRepository) : IRequestHandler<DeletePublisherCommand, Result>
 {
-    public async Task Handle(DeletePublisherCommand request, CancellationToken cancellationToken)
+    public async Task<Result> Handle(DeletePublisherCommand request, CancellationToken cancellationToken)
     {
-        _ = await publisherRepository.GetByIdAsync(request.Id, cancellationToken) ??
-        throw new NotFoundWithTheIdException(typeof(Publisher), request.Id);
+        Publisher? publisher = await publisherRepository.GetByIdAsync(request.Id, cancellationToken);
+        if (publisher is null)
+            return Result.Failure(PublisherAggregateErrors.NotFound(request.Id));
 
         //TODO: var hasPublication = (await publicationRepository.ListAllAsync(x => x.Publisher))
         await publisherRepository.DeleteAsync(request.Id, cancellationToken);
+        return Result.Success();
     }
 }
